Created deploy task to push or update application
Push & Update get a duplicate free list of URIs to bind to.
Update will bind to pre definded services if available.
Updated readme to include a section use of cf-deploy task, with example
of creating services before execution of the cf-deploy task.
Adding a service e.g. calling cf-add-service, is now an idempotent
operation. Meaning if you already have the service it won't create it
again or blow up because it already exists.